v0.0.9.8.1: AGC + persistence patch
Substantial bug-fix patch over v0.0.9.8.  Headline: a latent
SetRXAAGCSlope cffi binding bug present since v0.0.9.6 was
caught by an audit of every binding's parameter types vs the
WDSP C source.  Fixed.  AGC profiles (Fast/Med/Slow/Long) now
sound audibly distinct on real signals for the first time
since v0.0.9.6.

AGC bug-hunt:
- SetRXAAGCSlope cffi binding "double" -> "int" matches the
  C source; was producing register-class calling-convention
  garbage on Windows x86_64 (cffi pushed via XMM1 / double,
  C function read RDX / int).  var_gain ended up at a random
  value at every channel open, max_gain therefore wonky, and
  AGC pinned at whatever max_gain landed on regardless of
  profile.
- _open_wdsp_rx field-name fix — was reading
  self._agc_threshold which doesn't exist (legacy field was
  renamed to self._agc_target during the v0.0.9.6 cleanup
  arc), AttributeError caught by try/except, threshold push
  silently skipped at every channel open.
- set_agc_threshold now actually pushes to WDSP (was a
  store-and-emit-signal-only no-op).
- _agc_target semantic switched from legacy 0..1 linear
  output-target to dBFS-domain WDSP threshold value.  Default
  -100 dBFS gives ~70 dB AGC headroom (industry-typical).
- auto_set_agc_threshold reads live FFT noise floor
  (_noise_floor_db) instead of hardcoded _noise_baseline =
  1e-4 (the legacy noise-floor tracker that wrote
  _noise_baseline was deleted in v0.0.9.6 cleanup but the
  consumer was orphaned — Auto threshold has been static at
  -75 dBFS regardless of band conditions ever since).
  Margin lowered 18 -> 5 dB.
- AGC slope default 0 -> 35 (var_gain 1.0 -> 1.5 = soft-knee
  matching WDSP / industry convention).
- Settings -> DSP -> AGC threshold slider replaced with label
  readout + Auto button (the legacy 0..1 linear slider was
  advisory-only since v0.0.9.6; clean UX win).
- Long AGC profile re-added to the right-click menu —
  full radio-side wiring already present since v0.0.9.6, only
  the UI exposure was missing.

Per-band scale persistence repair:
- Operator-reported "Waterfall Min-Max isn't staying where
  set when restarting" + "spectrum floor not saved either".
- Two-part bug:
  1. _apply_band_range only fired from recall_band (band
     button click), never at startup — restored values were
     correctly in _band_memory but never APPLIED to live state.
  2. Spectrum autoload called set_spectrum_db_range with
     default from_user=True, which triggered
     _save_current_band_range to write CURRENT _waterfall_*
     fields (still constructor defaults at that load point)
     into _band_memory[band], CLOBBERING the just-restored
     operator values.  Same problem hit range_*, locks
     because all six fields are saved together.
  3. Auto-scale tick overwrote live _waterfall_* values 2
     sec after startup even when band had a manually-set
     waterfall in band_memory.
- Fixes: new Radio.apply_current_band_range() called from
  app.py after autoload completes; spectrum autoload now uses
  from_user=False; auto-scale skips waterfall mirror when
  band has manual waterfall in band_memory.

RX bandwidth persistence:
- _rx_bw_by_mode (per-mode RX bandwidth dict) was NEVER
  saved/loaded — every restart silently reset to BW_DEFAULTS.
  Caught alongside the waterfall persistence investigation.
  Now JSON-serialized to QSettings under "rx_bw_by_mode".

Click-to-tune snap refinement:
- _snap_tune_min_snr_db raised 6 -> 8 dB (less noise-blip
  false positives).
- New SNAP_RANGE_MAX_HZ = 2000 cap on effective search
  radius.  At 192 kHz / 1500 px the SNAP_PIXEL_RADIUS = 80
  px translated to ~10 kHz effective range — way too wide;
  cap keeps snap "the signal you pointed at" at every zoom
  level.

Documentation:
- CLAUDE.md §15.1 (current-version line, §14.2 wired/inert
  lists, "Last updated" trailer) — CLOSED.
- CLAUDE.md §15.5 (Long re-add) — CLOSED.
- CLAUDE.md §9.8 (Speaker-selective audio attenuator) —
  WITHDRAWN per operator (post-WDSP audio chain covers the
  use cases).
- AGC docs (agc.md, index.md, troubleshooting.md) — Long
  restored.

+++ lyra/dsp/wdsp_native.py
@@ -205,7 +205,16 @@ void SetRXAAGCDecay(int channel, int decay_ms);
 void SetRXAAGCHang(int channel, int hang_ms);
 void SetRXAAGCTop(int channel, double max_gain_db);
 void SetRXAAGCThresh(int channel, double thresh_db, double size, double rate);
-void SetRXAAGCSlope(int channel, double slope);
+/* SetRXAAGCSlope: WDSP source declares the slope parameter as
+ * `int slope` (units: 0.1 dB; passed via the integer calling
+ * convention).  Earlier Lyra cffi declarations had `double slope`
+ * which on Windows x86_64 caused a register-class mismatch — cffi
+ * passed the value via XMM1 (double) but the C function reads
+ * RDX (int), getting register garbage.  Result: var_gain (and
+ * therefore max_gain) was set to a random value at every channel
+ * open, which presented as "AGC pinned, no profile differences"
+ * since v0.0.9.6.  Fixed v0.0.9.8 to match the C source. */
+void SetRXAAGCSlope(int channel, int slope);
 void SetRXAAGCHangThreshold(int channel, int hangthreshold);
 
 /* ----- Meters (meter.c) ------------------------------------------------ */
